+++ packages/ui/src/components/popcorn/emissions-dashboard/ContractContainer/index.stories.tsx
@@ -1,7 +1,7 @@
 import { CloudIcon, TrendingUpIcon } from '@heroicons/react/outline';
 import { Meta, Story } from '@storybook/react/types-6-0';
 import React from 'react';
-import { EmissionSummaryStats } from '../../../interfaces/index';
+import { EmissionSummaryStats } from '../../../../interfaces/index';
 import { ContractContainer } from './index';
 
 const emissionSummaryStats: EmissionSummaryStats[] = [
