+++ packages/ui/src/components/popcorn/emissions-dashboard/recharts/AreaBarChart/index.tsx
