+++ packages/ui/src/components/popcorn/index.tsx
@@ -1,9 +1,8 @@
-export * from './AddContractButton';
-export * from './ContractContainer';
-export * from './DateRangePicker';
-export * from './Divider';
-export * from './NavBar';
-export * from './recharts/AreaBarChart/';
-export * from './StatsRow';
+export * from './emissions-dashboard/ContractContainer';
+export * from './emissions-dashboard/DateRangePicker';
+export * from './emissions-dashboard/Divider';
+export * from './emissions-dashboard/NavBar';
+export * from './emissions-dashboard/recharts/AreaBarChart/';
+export * from './emissions-dashboard/StatsRow';
+export * from './emissions-dashboard/TotalsContainer';
 export * from './Tooltip';
-export * from './TotalsContainer';
